# openapi.yaml
openapi: 3.0.0
info:
  title: My API
  version: 1.0.0
tags:
  - name: Users
    description: Operations related to users
servers:
  - url: http://localhost:3000
    description: Local server
paths:
  /api/users:
    get:
      summary: Get all users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                      example: John Doe
    post:
      summary: Create a new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: The created user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
  /api/users/{id}:
    put:
      summary: Update a user
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
    delete:
      summary: Delete a user
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Deletion message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /api/products:
    get:
      summary: Get all products
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                      example: Laptop
    post:
      summary: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: The created product
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
  /api/products/{id}:
    put:
      summary: Update a product
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: The updated product
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
    delete:
      summary: Delete a product
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Deletion message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /api/orders:
    get:
      summary: Get all orders
      responses:
        '200':
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    product:
                      type: string
                      example: Laptop
                    user:
                      type: string
                      example: John Doe
    post:
      summary: Create a new order
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: string
                user:
                  type: string
      responses:
        '200':
          description: The created order
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  product:
                    type: string
                  user:
                    type: string
  /api/orders/{id}:
    put:
      summary: Update an order
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: string
                user:
                  type: string
      responses:
        '200':
          description: The updated order
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  product:
                    type: string
                  user:
                    type: string
    delete:
      summary: Delete an order
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Deletion message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /api/reviews:
    get:
      summary: Get all reviews
      responses:
        '200':
          description: A list of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    product:
                      type: string
                      example: Laptop
                    review:
                      type: string
                      example: Great product!
    post:
      summary: Create a new review
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: string
                review:
                  type: string
      responses:
        '200':
          description: The created review
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  product:
                    type: string
                  review:
                    type: string
  /api/reviews/{id}:
    put:
      summary: Update a review
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: string
                review:
                  type: string
      responses:
        '200':
          description: The updated review
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  product:
                    type: string
                  review:
                    type: string
    delete:
      summary: Delete a review
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Deletion message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string